% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_to_edge.R
\name{adj_to_edge}
\alias{adj_to_edge}
\title{Convert an adjacency matrix to an edge list}
\usage{
adj_to_edge(adj_mat)
}
\arguments{
\item{adj_mat}{an adjacency matrix that shall be transformed to an edge list.}
}
\description{
This function takes a weighted adjacency matrix and
transforms it into an edge list
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
