% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_adj_edges.R
\name{adj_to_edges}
\alias{adj_to_edges}
\title{Convert an adjacency matrix to an edge list}
\usage{
adj_to_edges(adj)
}
\arguments{
\item{adj}{the adjacency matrix}
}
\description{
This function takes an adjacency matrix and
transforms it into an edge list
}
\author{
Ilias Moutsopoulos
}
