% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beeline_generate.R
\name{beeline_generate}
\alias{beeline_generate}
\title{Beeline input generation}
\usage{
beeline_generate(file_path, time_points = 2, samples = 2, column1names = TRUE)
}
\arguments{
\item{file_path}{a file path to an expression count matrix.}
}
\description{
This function writes the necessary files required to make a full
run of Beeline
}
\examples{
# path to a valid count matrix file
count_matrix_address <- system.file('beeline_examples/count_matrix/1000_raw_counts.csv', package = 'edgynode')
# import the count matrix
count_matrix <- read.csv(file = count_matrix_address)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
