% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beeline_tool_collection.R
\name{beeline_tool_collection}
\alias{beeline_tool_collection}
\title{Names of all individual BEELINE tools}
\usage{
beeline_tool_collection()
}
\description{
This is a list of all individual BEELINE tools for which output folders can be created.
}
