% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_adj_to_igraph.R
\name{convert_adj_to_igraph}
\alias{convert_adj_to_igraph}
\title{Convert an adjacency matrix to igraph object}
\usage{
convert_adj_to_igraph(adj)
}
\arguments{
\item{adj}{the adjacency matrix}
}
\description{
This function takes an adjacency matrix and
transforms it into an igraph object
}
\author{
Ilias Moutsopoulos and Sergio Vasquez
}
