% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_adj_to_matrix.R
\name{convert_adj_to_matrix}
\alias{convert_adj_to_matrix}
\title{Convert an adjacency matrix to plain matrix}
\usage{
convert_adj_to_matrix(adj)
}
\arguments{
\item{adj}{the adjacency matrix}
}
\description{
This function takes an adjacency matrix and
transforms it into a plain matrix
}
\author{
Ilias Moutsopoulos and Sergio Vasquez
}
