% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_edges_to_adj.R
\name{convert_edges_to_adj}
\alias{convert_edges_to_adj}
\title{Convert an edge list to an adjacency matrix}
\usage{
convert_edges_to_adj(edges)
}
\arguments{
\item{edges}{the edge list data frame, with 2 or 3 columns
(from, to, value)}
}
\description{
This function takes an edge list and
transforms it into an adjacency matrix
}
\author{
Ilias Moutsopoulos
}
