% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_import_genie.R
\name{genie}
\alias{genie}
\title{Output Parser for GENIE3}
\usage{
genie(file_path)
}
\arguments{
\item{file_path}{a file path to the GENIE3 output file generated by BEELINE.}
}
\description{
This function parses the standard output generated by the BEELINE tool GENIE3.
}
\examples{
# path to GENIE3 output file
genie3_output <- system.file('beeline_examples/GENIE3/outFile.csv', package = 'edgynode')
# import GENIE3 specific output
genie3_parsed <- genie(genie3_output)
# look at output
head(genie3_parsed)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
