% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_import_grisli.R
\name{grisli}
\alias{grisli}
\title{Output Parser for GRISLI}
\usage{
grisli(file_path)
}
\arguments{
\item{file_path}{a file path to the GRISLI output file generated by BEELINE.}
}
\description{
This function parses the standard output generated by the BEELINE tool GRISLI.
}
\examples{
# path to GRISLI output file
genie3_output <- system.file('beeline_examples/GENIE3/outFile.csv', package = 'scNetworkR')
# import GENIE3 specific output
genie3_parsed <- genie(genie3_output)
# look at output
head(genie3_parsed)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
