% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_genie3.R
\name{import_genie3}
\alias{import_genie3}
\title{Output Parser for GENIE3}
\usage{
import_genie3(filename)
}
\arguments{
\item{filename}{a file path to the GENIE3 output file generated by BEELINE.}
}
\description{
This function parses the standard output generated by the BEELINE tool GENIE3.
}
\examples{
# path to GENIE3 output file
genie3_output <- system.file('beeline_examples/GENIE3/outFile.csv', package = 'edgynode')
# import GENIE3 specific output
genie3_parsed <- import_genie3(genie3_output)
# look at output
head(genie3_parsed)
}
\author{
Ilias Moutsopoulos
}
