% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_adjacency.R
\name{is_adjacency}
\alias{is_adjacency}
\title{Check whether a \code{\link{make_adjacency}} converted matrix fulfills  the \pkg{edgynode} adjacency matrix convention}
\usage{
is_adjacency(x)
}
\arguments{
\item{x}{a \code{\link{matrix}} object.}
}
\description{
Imports a matrix and tests whether it is an adjacency in \pkg{edgynode} format.
}
\examples{
# test with an already clean adjacency matrix
edgynode::is_adjacency(edgynode::adjacency_clean_test_3)

# look at raw matrix
edgynode::adjacency_matrix_test_3
# convert raw matrix into a edgynode adjacency matrix
clean_matrix <- edgynode::make_adjacency(edgynode::adjacency_matrix_test_3)
# test converted matrix
edgynode::is_adjacency(clean_matrix)
}
\seealso{
\code{\link{is_symmetric}}, \code{\link{is_standard}}, \code{\link{is_binary}}
}
\author{
Ilias Moutsopoulos and Hajk-Georg Drost
}
