% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_binary.R
\name{is_binary}
\alias{is_binary}
\title{Check if a numeric matrix generated with \code{\link{make_adjacency}} and \code{\link{make_binary}} is binary}
\usage{
is_binary(adj)
}
\arguments{
\item{adj}{an adjacency matrix converted from a raw input matrix via \code{\link{make_adjacency}} and optionally binarised with \code{\link{make_binary}}.}
}
\description{
This function takes an adjacency matrix as input and
checks if it is binary
}
\author{
Ilias Moutsopoulos
}
