% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_standard.R
\name{is_standard}
\alias{is_standard}
\title{Check if a matrix is in a standard format}
\usage{
is_standard(adj, max_value = 1, no_negative = TRUE, no_self_loops = TRUE)
}
\arguments{
\item{adj}{the adjacency matrix}

\item{max_value}{the maximum (positive) value that should exist in the
matrix; default is 1; setting this to \code{NULL} or \code{max(abs(adj))}
leaves the values unchanged}

\item{no_negative}{whether there should be negative values or just their
absolute value (e.g. in the case of correlations); default is TRUE}

\item{no_self_loops}{whether the values in the main diagonal of the
adjacency matrix tshould be 0; default is TRUE}
}
\description{
This function takes an adjacency matrix as input and checks
if it is in a standard form (based on the options provided).
}
\author{
Ilias Moutsopoulos
}
