% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard}
\alias{jaccard}
\title{Compute a Jaccard/Tanimoto similarity coefficient between two vectors}
\usage{
jaccard(x, y, center = FALSE, px = NULL, py = NULL)
}
\arguments{
\item{x}{a binary vector}

\item{y}{a binary vector}

\item{center}{whether to center the Jaccard/Tanimoto coefficient by its expectation}

\item{px}{probability of successes in \code{x} (optional)}

\item{py}{probability of successes in \code{y} (optional)}
}
\description{
Compute a Jaccard/Tanimoto similarity coefficient between two vectors
}
\examples{
set.seed(537)
x = rbinom(50,1,.5)
y = rbinom(50,1,.4)
jaccard(x,y)
}
\references{
The initial implementation was taken from the \pkg{jaccard} package
by Neo Christopher Chung, Błażej Miasojedow, Michał Startek, Anna Gambin which is 
currently not availabe for R version >4.0 due to a unmaintained package dependency \pkg{qvalue}
and thus the source code of the function \code{jaccard} was modified for use in this package.
}
\author{
Hajk-Georg Drost
}
