% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_import_leap.r
\name{leap}
\alias{leap}
\title{Output Parser for LEAP}
\usage{
leap(file_path)
}
\arguments{
\item{file_path}{a file path to the LEAP output file generated by BEELINE.}
}
\description{
This function parses the standard output generated by the BEELINE tool LEAP.
}
\examples{
# path to LEAP output file
leap_output <- system.file('beeline_examples/LEAP/outFile0.txt', package = 'edgynode')
# import LEAP specific output
leap_parsed <- leap(leap_output)
# look at output
head(leap_parsed)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
