% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adjacency.R
\name{make_adjacency}
\alias{make_adjacency}
\title{Convert a raw input matrix into an \pkg{edgynode} adjacency matrix}
\usage{
make_adjacency(x, known_binary = FALSE)
}
\arguments{
\item{x}{a \code{\link{matrix}} object.}

\item{known_binary}{a logical value indicating whether the input matrix
\code{x} is already known to be a binary matrix
(for this case use \code{known_binary = TRUE}) or not
(for this case use \code{known_binary = FALSE} (default)).}

\item{known_symmetric}{a logical value indicating whether the input matrix
\code{x} is already known to be symmetric
(for this case use \code{known_symmetric = TRUE}) or not
(for this case use \code{known_symmetric = FALSE} (default)).}
}
\value{
The input matrix, converted to an adjacency
}
\description{
Users may work with a vast variety of input matrices that
are derived from various gene regulatory network inference methods and tools
with various values and value ranges. This function aims to ensure that
such raw data is converted into a \pkg{edgynode} adjacency matrix format
and is the first conversion step when importing existing \code{raw}
adjacency matrices into the \pkg{edgynode} ecosystem.
}
\examples{
# look at raw matrix
edgynode::adjacency_matrix_test_3
# convert raw matrix into a edgynode adjacency matrix
edgynode::make_adjacency(edgynode::adjacency_matrix_test_3)
}
\seealso{
\code{\link{assert_adjacency}}, \code{\link{is_adjacency}}
}
