% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_binary.R
\name{make_binary}
\alias{make_binary}
\title{Convert a numeric matrix to binary}
\usage{
make_binary(adj, threshold = 0.7)
}
\arguments{
\item{adj}{the adjacency matrix}

\item{threshold}{the threshold used for binarisation; default is \code{0.5}}
}
\description{
This function takes an adjacency matrix as input and
converts it to a binary one
}
\author{
Ilias Moutsopoulos
}
