% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_standard.R
\name{make_standard}
\alias{make_standard}
\title{Force a numeric matrix generated with \code{\link{make_adjacency}} into the \pkg{edgynode} standard matrix format}
\usage{
make_standard(
  adj,
  max_value = 1,
  split_by_margin = NULL,
  no_negative = TRUE,
  no_self_loops = TRUE
)
}
\arguments{
\item{adj}{an adjacency matrix converted from a raw input matrix via \code{\link{make_adjacency}}.}

\item{max_value}{the maximum (positive) value that should exist in the
matrix; default is 1; setting this to \code{NULL} or \code{max(abs(adj))}
leaves the values unchanged}

\item{no_negative}{whether there should be negative values or just their
absolute value (e.g. in the case of correlations); default is \code{no_negative = TRUE}.}

\item{no_self_loops}{whether the values in the main diagonal of the
adjacency matrix should be 0; default is \code{no_negative = TRUE}.}
}
\description{
This function takes an adjacency matrix as input and converts
it into a standard form (based on the options provided).
}
\seealso{
\code{\link{make_adjacency}}, \code{\link{make_symmetric}}, \code{\link{make_binary}}
}
\author{
Ilias Moutsopoulos
}
