% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_symmetric.R
\name{make_symmetric}
\alias{make_symmetric}
\title{Force a matrix into being symmetric}
\usage{
make_symmetric(adj, method = pmax)
}
\arguments{
\item{adj}{the adjacency matrix}

\item{method}{the element-wise function to be applied to the matrix and
its transpose; default is \code{pmax}}
}
\description{
This function takes an adjacency matrix as input and forces it
into being symmetric.
}
\author{
Ilias Moutsopoulos
}
