% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_benchmark_noise_filtering_test.R
\name{network_benchmark_noise_filtering_test}
\alias{network_benchmark_noise_filtering_test}
\title{Perform a statistical test on noise filter benchmarking results}
\usage{
network_benchmark_noise_filtering_test(
  network_benchmark_noise_filtering_result,
  test_type = NULL
)
}
\arguments{
\item{network_benchmark_noise_filtering_result}{a benchmark result \code{tibble} generated with \code{\link{network_benchmark_noise_filtering}}}

\item{test_type}{option to assess the statistical significance between comparison differences. Default is \code{test_type = NULL}.
\itemize{
\item \code{test_type = "kruskal"}: performs the Kruskal-Wallis Rank Sum Test to assess the statistical signifiance between comparison differences.
}}
}
\description{
This function takes a noise-filtering benchmark result generated with
\code{\link{network_benchmark_noise_filtering}} as input and performs a statistical test
to assess the statistical significance between comparison differences.
}
\examples{
genie3_49_raw <- as.matrix(read.csv(
system.file("data/network_raw_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_noNoiseCM_raw <- as.matrix(read.csv(
system.file("data/network_noNoiseCM_raw_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_qnorm_no_noise_removed <- as.matrix(read.csv(
system.file("data/network_qnorm_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_noNoiseCM_qnorm <- as.matrix(read.csv(
system.file("data/network_noNoiseCM_qnorm_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

# Run Benchmark using Hamming distance
benchmark_hamming <- network_benchmark_noise_filtering(
genie3_49_raw,
genie3_49_noNoiseCM_raw,
genie3_49_qnorm_no_noise_removed,
genie3_49_noNoiseCM_qnorm,
dist_type = "hamming",
grn_tool = "GENIE3")

# visualize at results
plot_network_benchmark_noise_filtering(benchmark_hamming,
            dist_type = "hamming", 
            title = "Network Inference Tool: GENIE3")
            
# perform Kruskal-Wallis Rank Sum Test
network_benchmark_noise_filtering_test(benchmark_hamming, test_type = "kruskal")
}
\author{
Hajk-Georg Drost
}
