% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_dist_pairwise_genes.R
\name{network_dist_pairwise_genes}
\alias{network_dist_pairwise_genes}
\title{Comparison of binary adjacency weights of the same gene in two different networks}
\usage{
network_dist_pairwise_genes(
  adj_mat_qry,
  adj_mat_sbj,
  dist_type = "hamming",
  print_message = TRUE
)
}
\arguments{
\item{adj_mat_qry}{a binary adjacency matrix generated with \code{\link{network_rescale}} and \code{\link{network_make_binary}}.}

\item{adj_mat_sbj}{a binary adjacency matrix with the same genes as in \code{adj_mat_qry}, but with differnt binary edge weights generated with \code{\link{network_rescale}} and \code{\link{network_make_binary}}.}

\item{dist_type}{a distance method that shall be applied on the binary values for each gene. Available options are:
\itemize{
\item \code{dist_type = "hamming"}: computes the \code{\link[e1071]{hamming.distance}} for each gene between the two input matrices
\item \code{dist_type = "jaccard"}: computes the \code{\link[jaccard]{jaccard}} for each gene between the two input matrices
}}

\item{print_message}{shall massages be printed? Default is \code{print_message = TRUE}.}
}
\description{
Compute the pairwise \code{Hamming} or \code{Jaccard} distance between binary edge weights of the 
same genes in two different network topologies (denoted as \code{adj_mat_qry} and \code{adj_mat_sbj}).
}
\examples{
## Import and rescale PIDC network
# path to PIDC output file
pidc_output <- system.file('beeline_examples/PIDC/outFile.txt', package = 'edgynode')
# import PIDC specific output
pidc_parsed <- pidc(pidc_output)
#Set diagonal values
diag(pidc_parsed) <- 1
# rescaling PIDC output
pidc_rescaled <- network_rescale(pidc_parsed)
pidc_binary_adj_mat <- network_make_binary(pidc_rescaled, threshold = "median")
# compute hamming distances for each gene between input matrices
network_dist_pairwise_genes(pidc_binary_adj_mat, 
pidc_binary_adj_mat, dist_type = "hamming")

# compute jaccard distances for each gene between input matrices
network_dist_pairwise_genes(pidc_binary_adj_mat, 
pidc_binary_adj_mat, dist_type = "jaccard")
}
\author{
Hajk-Georg Drost
}
