% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_filter_edges.R
\name{network_filter_edges}
\alias{network_filter_edges}
\title{Matrix filter edges}
\usage{
network_filter_edges(
  adj_mat,
  threshold = NULL,
  print_message = TRUE,
  p_value = 0.1
)
}
\arguments{
\item{adj_mat}{adjacency matrix to be converted.}

\item{threshold}{if not specified, the function will filter based on the 
disparity measure (Neal, Z.P. (2022). backbone: An R package to extract 
network backbones) (recommended). If you want to specify threshold, we recommended to use \code{\link{network_rescale}}
before using this function. Re-scaling will transform all values into a range [0,100].
The threshold can either be a numeric balue in the interval [0,100] or a character string
specifying the following methods for automatically determining the threshold based on the input data:}

\item{print_message}{a logical value indicating whether or not a threshold message shall be printed.}

\item{p_value}{a value between setting the treshold for hypothesis testing for the backbone filtering.}
}
\description{
This function takes a weighted adjacency matrix filters out
the unsignificant edges
}
\examples{
 # path to PPCOR output file
ppcor_output <- system.file('beeline_examples/PPCOR/outFile.txt', package = 'edgynode')
 # import PPCOR specific output
ppcor_parsed <- ppcor(ppcor_output)
 # rescale the matrix
ppcor_rescaled <- network_rescale(ppcor_parsed)
 # make the binary matrix
ppcor_binary <- network_make_binary(ppcor_rescaled)
 # to see the differences you may use:
 plot_network_dendrogram(ppcor_rescaled)
 plot_network_dendrogram(ppcor_binary)
}
\author{
Nikola Kalábová, Sergio Vasquez and Hajk-Georg Drost
}
