% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_make_binary.R
\name{network_make_binary}
\alias{network_make_binary}
\title{Matrix convert to binary}
\usage{
network_make_binary(adj_mat, threshold, print_message = TRUE)
}
\arguments{
\item{adj_mat}{adjacency matrix to be converted.}

\item{threshold}{we recommended to use \code{\link{network_rescale}}
before using this function. Re-scaling will transform all values into a range [0,100].}

\item{print_message}{a logical value indicating whether or not a threshold message shall be printed.}
}
\description{
This function takes a weighted adjacency matrix and separates
the components of the matrix into weights and presence/absence matrices
making use of a threshold.
}
\examples{
 # path to PPCOR output file
ppcor_output <- system.file('beeline_examples/PPCOR/outFile.txt', package = 'edgynode')
 # import PPCOR specific output
ppcor_parsed <- ppcor(ppcor_output)
 # rescale the matrix
ppcor_rescaled <- network_rescale(ppcor_parsed)
 # make the binary matrix
ppcor_binary <- network_make_binary(ppcor_rescaled, 70)
 # to see the differences you may use:
 plot_network_dendrogram(ppcor_rescaled)
 plot_network_dendrogram(ppcor_binary)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
