% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_symmetric.r
\name{network_make_symmetric}
\alias{network_make_symmetric}
\title{Force a matrix into being symmetric}
\usage{
network_make_symmetric(adj_mat)
}
\arguments{
\item{adj_mat}{an adjacency matrix that shall be made symmetric.}
}
\description{
This function takes and \code{adj_mat} as input and forces it into being symmetric.
}
\author{
lmshk
}
