% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot_dendrogram.R
\name{network_plot_dendrogram}
\alias{network_plot_dendrogram}
\title{Function to plot a dendrogram based on gene interaction}
\usage{
network_plot_dendrogram(adj_mat, dist_method = "euclidean")
}
\arguments{
\item{adj_mat}{an adjacency matrix to plot the dendrogram from.}

\item{dist_method}{method to be used to calculate distances
Default is \code{euclidean}.}
}
\description{
This function takes an adjacency matrix representing a
gene regulatory network and builds a dendrogram based on
the stength of gene interactions
}
\examples{
# path to PIDC output file
pidc_output <- system.file('beeline_examples/PIDC/outFile.txt', package = 'scNetworkR')
# parsing the output to an adjacency matrix
pidc_parsed <- pidc(pidc_output)
# rescaling the matrix
rescaled <- network_rescale(pidc_parsed)
# Visualize result
network_plot_dendrogram(rescaled)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
