% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_rescale.R
\name{network_rescale}
\alias{network_rescale}
\title{Rescaling function customized for adjacency matrices}
\usage{
network_rescale(adj_mat, to = c(0, 100))
}
\arguments{
\item{adj_mat}{an adjacency matrix to normalize.}

\item{to}{interval in which transformation should be done.
Default is \code{to = c(0, 100)}.}
}
\description{
This function rescales to a 0 to 1 scale using the values
present in the matrix. It is based on the rescale function of the
scales package.
}
\examples{
# path to GENIE3 output file
genie3_output <- system.file('beeline_examples/GENIE3/outFile.csv', package = 'edgynode')
# parsing the output to an adjacency matrix
genie3_parsed <- genie(genie3_output)
# rescaling the matrix
rescaled <- network_rescale(genie3_parsed)
# Visualize result
head(rescaled)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
