% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_statistics_degree_distribution_naive.R
\name{network_statistics_degree_distribution_naive}
\alias{network_statistics_degree_distribution_naive}
\title{Calculate the naive degree distribution from binary adjacency matrices}
\usage{
network_statistics_degree_distribution_naive(adj_mat)
}
\arguments{
\item{adj_mat}{a symmetrical binary adjacency matrix.}
}
\description{
This function calculates
the degree distribution from a binary adjacency matrix.
}
\examples{
# path to PPCOR output file
ppcor_output <- system.file('beeline_examples/PPCOR/outFile.txt', package = 'edgynode')
# import PPCOR output into adjacency matrix
ppcor_parsed <- ppcor(ppcor_output)
# rescaling matrix
ppcor_rescaled <- network_rescale(ppcor_parsed)
# make weighted adjacency matrix binary
ppcor_binary_adj_mat <- network_make_binary(ppcor_rescaled, threshold = 70)
# calculate network statistics
ppcor_statistics <- network_statistics_degree_distribution_naive(ppcor_binary_adj_mat)
# look at results
ppcor_statistics
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
