% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_import_pidc.R
\name{pidc}
\alias{pidc}
\title{Output Parser for PIDC}
\usage{
pidc(file_path)
}
\arguments{
\item{file_path}{a file path to the PIDC output file generated by BEELINE.}
}
\description{
This function parses the standard output generated by the BEELINE tool PIDC.
}
\examples{
# path to PIDC output file
pidc_output <- system.file('beeline_examples/PIDC/outFile.txt', package = 'edgynode')
# import PIDC specific output
pidc_parsed <- pidc(pidc_output)
# look at output
head(pidc_parsed)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
