% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_benchmark_noise_filtering.R
\name{plot_network_benchmark_noise_filtering}
\alias{plot_network_benchmark_noise_filtering}
\title{Visualize noise filter benchmarking results}
\usage{
plot_network_benchmark_noise_filtering(
  network_benchmark_noise_filtering_result,
  dist_type,
  xlab = "Comparison between original, filtered, and normalized networks",
  title = "",
  y_ticks = 10
)
}
\arguments{
\item{network_benchmark_noise_filtering_result}{a benchmark result \code{tibble} generated with \code{\link{network_benchmark_noise_filtering}}.}

\item{dist_type}{distance method that shall be applied on the binary values for each gene. Available options are:
\itemize{
\item \code{dist_type = "hamming"}: computes the \code{\link[e1071]{hamming.distance}} for each gene between the two input matrices
\item \code{dist_type = "jaccard"}: computes the \code{\link[jaccard]{jaccard}} for each gene between the two input matrices
}
See \code{\link{network_dist_pairwise_genes}} for details.

Finally, a \code{\link{kruskal.test}} is performed to assess the statistical significance of comparison differences.}

\item{xlab}{x-axis label.}

\item{title}{a character string denoting the plot title.}

\item{y_ticks}{the number of ticks that shall be drawn on the y-axis.}
}
\description{
This function takes a noise-filtering benchmark result generated with
\code{\link{network_benchmark_noise_filtering}} as input and visualizes the results as violin-plots.
}
\examples{
# Benchmark GENIE3 inferred networks with raw, no_noise, and quantile_norm combinations
genie3_49_raw <- as.matrix(read.csv(
system.file("data/network_raw_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_noNoiseCM_raw <- as.matrix(read.csv(
system.file("data/network_noNoiseCM_raw_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_qnorm_no_noise_removed <- as.matrix(read.csv(
system.file("data/network_qnorm_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_noNoiseCM_qnorm <- as.matrix(read.csv(
system.file("data/network_noNoiseCM_qnorm_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

# Run Benchmark using Hamming distance
benchmark_hamming <- network_benchmark_noise_filtering(
genie3_49_raw,
genie3_49_noNoiseCM_raw,
genie3_49_qnorm_no_noise_removed,
genie3_49_noNoiseCM_qnorm,
dist_type = "hamming",
grn_tool = "GENIE3")

# visualize at results
plot_network_benchmark_noise_filtering(benchmark_hamming,
            dist_type = "hamming", title = "Network Inference Tool: GENIE3")

}
\author{
Hajk-Georg Drost
}
