% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_degree_distribution_naive.R
\name{plot_network_degree_distribution_naive}
\alias{plot_network_degree_distribution_naive}
\title{Plot the naive degree distribution of a degree table}
\usage{
plot_network_degree_distribution_naive(
  degree_tbl,
  xlab = "Gene Name",
  ylab = "Node Degree per Gene",
  y_ticks = 12
)
}
\arguments{
\item{degree_tbl}{a degree table generated with
\code{\link{network_make_binary}} and \code{network_degree_distribution_naive}.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{y_ticks}{number of ticks on y-axis.}
}
\description{
This function takes a degree table generated with
\code{\link{network_make_binary}} and \code{network_degree_distribution_naive} as input
and visualizes the corresponding naive degree distribution.
}
\examples{
# path to PPCOR output file
ppcor_output <- system.file('beeline_examples/PPCOR/outFile.txt', package = 'edgynode')
# import PPCOR output into adjacency matrix
ppcor_parsed <- ppcor(ppcor_output)
# rescaling matrix
ppcor_rescaled <- network_rescale(ppcor_parsed)
# make weighted adjacency matrix binary
ppcor_binary_adj_mat <- network_make_binary(ppcor_rescaled, threshold = 70)
# calculate network statistics
ppcor_statistics <- network_statistics_degree_distribution_naive(ppcor_binary_adj_mat)
# look at results
plot_network_degree_distribution_naive(ppcor_statistics)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
