% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/plot_network_degree_distribution_naive_comparison.R
\name{plot_network_degree_distribution_naive_comparison}
\alias{plot_network_degree_distribution_naive_comparison}
\title{Benchmarking of noise filtering and normalization impact on network inference}
\usage{
plot_network_degree_distribution_naive_comparison(
  adj_mat_not_filtered_not_normalized,
  adj_mat_filtered_and_not_normalized,
  adj_mat_not_filtered_but_normalized,
  adj_mat_filtered_and_normalized,
  threshold = "median",
  xlab = "Gene",
  ylab = "Node degree",
  y_ticks = 15,
  print_message = TRUE
)
}
\arguments{
\item{adj_mat_not_filtered_not_normalized}{a weighted adjacency matrix derived from a network inference program
where no noise-filtering or quantile-normalization was applied to the input data.}

\item{adj_mat_filtered_and_not_normalized}{a weighted adjacency matrix derived from a network inference program
where noise-filtering but no quantile-normalization was applied to the input data.}

\item{adj_mat_not_filtered_but_normalized}{a weighted adjacency matrix derived from a network inference program
where no noise-filtering but quantile-normalization was applied to the input data.}

\item{adj_mat_filtered_and_normalized}{a weighted adjacency matrix derived from a network inference program
where noise-filtering and quantile-normalization were applied to the input data.}

\item{threshold}{we recommended to use \code{\link{network_rescale}}
before using this function. Re-scaling will transform all values into a range [0,100].
The threshold can either be a numeric value in the interval [0,100] or a character string
specifying the following methods for automatically determining the threshold based on the input data:
\itemize{
\item \code{threshold = "median"}: compute the \code{\link{median}} over the entire input \code{adj_mat} and use this
\code{median} value as threshold for defining all edge weights of a genes equal or below the \code{median}
value as \code{0} and all values above the \code{median} value as \code{1}.
}
See \code{\link{network_make_binary}} for details.
See \code{\link{network_statistics_degree_distribution_naive}} for details.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{y_ticks}{number of y-ticks.}

\item{print_message}{shall messages be printed?}
}
\description{
This function takes four types of matrices as input which
were generated with the same network inference tool (e.g. GENIE3), but had
different input data to the network inference tool in the context of
noise-filtering and quantile-normalization.
It is assumed that the four matrices inserted into this function come from
network inference runs with the following input data specifications:
\itemize{
\item Raw mapped RNAseq count data without noise-filtering or quantile-normalization applied (\code{adj_mat_not_filtered_not_normalized})
\item Raw mapped RNAseq count data with noise-filtering applied but no quantile-normalization (\code{adj_mat_filtered_and_not_normalized})
\item Raw mapped RNAseq count data without noise-filtering but with quantile-normalization applied (\code{adj_mat_not_filtered_but_normalized})
\item Raw mapped RNAseq count data with noise-filtering applied and with quantile-normalization applied (\code{adj_mat_filtered_and_normalized})
}

All relevant pairwise comparisons are then performed internally with \code{\link{network_dist_pairwise_genes}} based on gene-wise
Hamming Distance or Jaccard Similarity Coefficient computations.
}
\examples{
# Benchmark GENIE3 inferred networks with raw, no_noise, and quantile_norm combinations
genie3_49_raw <- as.matrix(read.csv(
system.file("data/network_raw_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_noNoiseCM_raw <- as.matrix(read.csv(
system.file("data/network_noNoiseCM_raw_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_qnorm_no_noise_removed <- as.matrix(read.csv(
system.file("data/network_qnorm_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))

genie3_49_noNoiseCM_qnorm <- as.matrix(read.csv(
system.file("data/network_noNoiseCM_qnorm_49_placenta_development.csv",
 package = "edgynode"), row.names = 1))


# Plot Results
plot_network_degree_distribution_naive_comparison(
genie3_49_raw,
genie3_49_noNoiseCM_raw,
genie3_49_qnorm_no_noise_removed,
genie3_49_noNoiseCM_qnorm)
}
\author{
lmshk and Hajk-Georg Drost
}
