% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_dendrogram.R
\name{plot_network_dendrogram}
\alias{plot_network_dendrogram}
\title{Function to plot a dendrogram based on gene interaction}
\usage{
plot_network_dendrogram(adj_mat, dist_method = "euclidean")
}
\arguments{
\item{adj_mat}{an adjacency matrix to plot the dendrogram from.}

\item{dist_method}{method to be used to calculate distances
Default is \code{euclidean}.}
}
\description{
This function takes an adjacency matrix representing a
gene regulatory network and builds a dendrogram based on
the stength of gene interactions
}
\examples{
# path to PPCOR output file
ppcor_output <- system.file('beeline_examples/PPCOR/outFile.txt', package = 'edgynode')
# import PPCOR specific output
ppcor_parsed <- ppcor(ppcor_output)
# rescale matrix
ppcor_rescaled <- network_rescale(ppcor_parsed)
# plot dendrogram
plot_network_dendrogram(ppcor_rescaled)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
