% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_weight_distribution.R
\name{plot_network_weight_distribution}
\alias{plot_network_weight_distribution}
\title{Generate correlation heatmap}
\usage{
plot_network_weight_distribution(
  adj_mat,
  xlab = "Edge weight",
  ylab = "Gene name",
  threshold = 70
)
}
\arguments{
\item{adj_mat}{adjacency matrix to be converted.}
}
\description{
This function takes a weighted adjacency matrix and generates 
a heatmap in which the highest values are moved towards the top left corner
to improve readability
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
