% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_weight_distribution_violin.R
\name{plot_network_weight_distribution_violin}
\alias{plot_network_weight_distribution_violin}
\title{Plot the weight distribution as violin plot}
\usage{
plot_network_weight_distribution_violin(
  adj_mat,
  xlab = "Edge weight",
  ylab = "Gene name",
  threshold
)
}
\arguments{
\item{adj_mat}{a weighted adjacency matrix.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{threshold}{we recommended to use \code{\link{network_rescale}}
before using this function. Re-scaling will transform all values into a range [0,100].
The threshold can either be a numeric balue in the interval [0,100] or a character string
specifying the following methods for automatically determining the threshold based on the input data:
\itemize{
\item \code{threshold = "median"}: compute the \code{\link{median}} over the entire input \code{adj_mat} and use this 
\code{median} value as threshold for defining all edge weights of a genes equal or below the \code{median}
value as \code{0} and all values above the \code{median} value as \code{1}.  
}
The threshold value will then be drawn as vertical line in the plot.}
}
\description{
This function takes a weighted adjacency matrix and plots
the weight distribution for all edges of a gene.
}
\examples{
# path to PPCOR output file
ppcor_output <- system.file('beeline_examples/PPCOR/outFile.txt', package = 'edgynode')
# import PPCOR output into adjacency matrix
ppcor_parsed <- ppcor(ppcor_output)
# rescaling weighted adjacency matrix to range [0,100]
ppcor_rescaled <- network_rescale(ppcor_parsed)
# visualize weight distributions over all edges for each gene
plot_network_weight_distribution_violin(ppcor_rescaled, threshold = 70)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
