% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_delta_con_similarity.R
\name{summary_delta_con_similarity}
\alias{summary_delta_con_similarity}
\title{Calculate DeltaCon similarity}
\usage{
summary_delta_con_similarity(adj, degrees, .MAX_POWER = 10, debug = FALSE)
}
\arguments{
\item{adj}{the adjacency matrix}

\item{degrees}{the vector of node degrees}

\item{.MAX_POWER}{maximum power for matrix inversion}

\item{debug}{if TRUE, the function will gives you the time it spent on each step}
}
\value{
The similarity matrix S as described in the DeltaCon method
}
\description{
Computes the similarity matrix of an adjacency based on the DeltaCon method.
Adapted from previous implementation by Baoxu(Dash) Shi, Data Sciense Group,
University of Notre Dame
}
\references{
D. Koutra, J. T. Vogelstein, and C. Faloutsos: DeltaCon:
A Principled Massive-Graph Similarity Function. SIAM 2013: 162-170.

D. Koutra, T. Ke, U. Kang, D. H. Chau, H. K. Pao, C. Faloutsos:
Unifying Guilt-by-Association Approaches: Theorems and Fast Algorithms.
ECML/PKDD (2) 2011: 245-260
}
