% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beeline_parser.R
\name{beeline_parser}
\alias{beeline_parser}
\title{General BEELINE Parser}
\usage{
beeline_parser(data_folder, name_tool)
}
\arguments{
\item{data_folder}{a file path to the BEELINE output folder storing all individual BEELINE tool runs (output folders).}

\item{name_tool}{name of the individual BEELINE tool for which output shall be imported (see tool collection with \code{\link{beeline_tool_collection}}).}
}
\description{
This function provides users with a wrapper function to parse the output files
of any BEELINE tool.
}
\examples{
# import BEELINE output file
beeline_out_folder <- system.file('beeline_examples', package = 'edgynode')
beeline_list <- beeline_parser(beeline_out_folder, beeline_tool_collection())
# look at example
beeline_list
}
\seealso{
\code{\link{network_import_genie}}, \code{\link{network_import_grisli}}
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
