% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_import_grisli.R
\name{grisli}
\alias{grisli}
\title{Output Parser for GRISLI}
\usage{
grisli(file_path)
}
\arguments{
\item{file_path}{a file path to the GRISLI output file generated by BEELINE.}
}
\description{
This function parses the standard output generated by the BEELINE tool GRISLI.
}
\examples{
# path to GRISLI output file
grisli_output <- system.file('beeline_examples/GRISLI/0/outFile.txt', package = 'edgynode')
# import GENIE3 specific output
grisli_parsed <- grisli(grisli_output)
# look at output
head(grisli_parsed)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
