% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_adjacency.R
\name{is_adjacency}
\alias{is_adjacency}
\title{Check whether input matrix is an adjacency matrix}
\usage{
is_adjacency(x)
}
\arguments{
\item{x}{a \code{\link{matrix}} object.}
}
\description{
Imports a matrix and tests whether it is an adjacency in \pkg{edgynode} format.
}
