% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_adjacency.R
\name{is_adjacency}
\alias{is_adjacency}
\title{Check whether a matrix fulfills  the \pkg{edgynode}
adjacency matrix convention}
\usage{
is_adjacency(x)
}
\arguments{
\item{x}{a \code{\link{matrix}} object.}
}
\value{
The result of the test (TRUE/FALSE)
}
\description{
Tests whether a matrix is an adjacency in \pkg{edgynode} format.
}
\examples{
# test with an already clean adjacency matrix
edgynode::is_adjacency(edgynode::adjacency_clean_test_3)
# look at raw matrix
edgynode::adjacency_matrix_test_3
# convert raw matrix into a edgynode adjacency matrix
clean_matrix <- edgynode::make_adjacency(edgynode::adjacency_matrix_test_3)
# test converted matrix
edgynode::is_adjacency(clean_matrix)
}
\seealso{
\code{\link{make_adjacency}}, \code{\link{assert_adjacency}}
}
\author{
Ilias Moutsopoulos and Hajk-Georg Drost
}
