% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_binary.R
\name{is_binary}
\alias{is_binary}
\title{Check if a matrix is binary}
\usage{
is_binary(adj)
}
\arguments{
\item{adj}{the adjacency matrix}
}
\description{
This function takes an adjacency matrix as input and
checks if it is binary
}
\author{
Ilias Moutsopoulos
}
