% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_symmetric.R
\name{is_symmetric}
\alias{is_symmetric}
\title{Check if a matrix is symmetric}
\usage{
is_symmetric(adj)
}
\arguments{
\item{adj}{the adjacency matrix}
}
\description{
This function takes an adjacency matrix as input and checks if
it is symmetric.
}
\author{
Ilias Moutsopoulos
}
