% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_make_binary.R
\name{network_make_binary}
\alias{network_make_binary}
\title{Matrix convert to binary}
\usage{
network_make_binary(adj_mat, threshold = 0.5)
}
\arguments{
\item{adj_mat}{adjacency matrix to be converted.}
}
\description{
This function takes a weighted adjacency matrix and separates
the components of the matrix into weights and presence/absence matrices
making use of a threshold.
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
