% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_make_binary.R
\name{network_make_binary}
\alias{network_make_binary}
\title{Matrix convert to binary}
\usage{
network_make_binary(adj_mat, threshold)
}
\arguments{
\item{adj_mat}{adjacency matrix to be converted.}

\item{threshold}{it's recommended to use the network_rescale() function
before using this function. This will require a threshold with value 0-100.}
}
\description{
This function takes a weighted adjacency matrix and separates
the components of the matrix into weights and presence/absence matrices
making use of a threshold.
}
\examples{
 # path to PPCOR output file
ppcor_output <- system.file('beeline_examples/PPCOR/outFile.txt', package = 'edgynode')
 # import PPCOR specific output
ppcor_parsed <- ppcor(ppcor_output)
 # rescale the matrix
ppcor_rescaled <- network_rescale(ppcor_parsed)
 # make the binary matrix
ppcor_binary <- network_make_binary(ppcor_rescaled, 70)
 # to see the differences you may use:
 network_plot_dendrogram(ppcor_rescaled)
 network_plot_dendrogram(ppcor_binary)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
