% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_rescale.R
\name{network_rescale}
\alias{network_rescale}
\title{Rescaling function customized for adjacency matrices}
\usage{
network_rescale(adj_mat, to = c(0, 100))
}
\arguments{
\item{adj_mat}{an adjacency matrix to normalize.}

\item{to}{interval in which transformation should be done.
Default is \code{to = c(0, 100)}.}
}
\description{
This function rescales to a 0 to 100 scale using the values
present in the matrix. It is based on the rescale function of the
scales package. It will make the matrix symmetric by mirroring the upper
half if it isn't. If negative values present, their absolute value will be
taken.
}
\examples{
# path to ppcor output file
ppcor_output <- system.file('beeline_examples/PPCOR/outFile.txt', package = 'edgynode')
# parsing the output to an adjacency matrix
ppcor_parsed <- ppcor(ppcor_output)
# rescaling the matrix
rescaled <- network_rescale(ppcor_parsed)
# Visualize result
head(rescaled)
}
\author{
Sergio Vasquez and Hajk-Georg Drost
}
